add fixed lab6

+++ fridge/src/main/java/ua/lviv/iot/algo/part1/lab1/controllers/WineFridgeController.java
@@ -1,5 +1,6 @@
 package ua.lviv.iot.algo.part1.lab1.controllers;
 
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -18,32 +19,45 @@ import java.util.Collection;
 public class WineFridgeController {
     private final WineFridgeService wineFridgeService;
 
-    public WineFridgeController(WineFridgeService wineFridgeService) {
+    public  WineFridgeController(WineFridgeService wineFridgeService) {
         this.wineFridgeService = wineFridgeService;
     }
 
     @GetMapping
-    public Collection<WineFridge> getAllFridges() {
-        return this.wineFridgeService.getAllFridges();
+    public ResponseEntity<Collection<WineFridge>> getAllFridges() {
+        return ResponseEntity.ok(this.wineFridgeService.getAllFridges());
     }
 
     @GetMapping("/:id")
-    public WineFridge getFridgeById(@PathVariable("id") Integer id) {
-        return this.wineFridgeService.getFridgeById(id);
+    public ResponseEntity<WineFridge> getFridgeById(@PathVariable("id") Integer id) {
+        WineFridge entity = this.wineFridgeService.getFridgeById(id);
+        if (entity == null){
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(entity);
     }
 
     @PostMapping
-    public WineFridge addFridge(@RequestBody WineFridge entity) {
-        return this.wineFridgeService.addFridge(entity);
+    public ResponseEntity<WineFridge> addFridge(@RequestBody WineFridge entity) {
+        return ResponseEntity.ok( this.wineFridgeService.addFridge(entity));
     }
 
     @PutMapping("/:id")
-    public WineFridge editFridge(@PathVariable Integer id, @RequestBody WineFridge entity) {
-        return this.wineFridgeService.editFridge(id, entity);
+    public ResponseEntity<WineFridge> editFridge(@PathVariable Integer id, @RequestBody WineFridge entity) {
+        WineFridge entity1 = this.wineFridgeService.editFridge(id, entity);
+        if (entity1 == null){
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(entity1);
     }
 
     @DeleteMapping("/:id")
-    public void deleteFridge(@PathVariable Integer id) {
+    public ResponseEntity<String> deleteFridge(@PathVariable Integer id) {
+        WineFridge entity1 = this.wineFridgeService.getFridgeById(id);
+        if (entity1 == null){
+            return ResponseEntity.notFound().build();
+        }
         this.wineFridgeService.deleteFridge(id);
+        return ResponseEntity.ok("fridge with id: " + id);
     }
 }
